backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.5
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
numpy==1.25.2
oauth2==1.9.0.post1
oauthlib==3.2.2
pandas==2.1.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-openid==2.2.5
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
